Refactor MenuModel components to it's own file and remove default imports
