+++ src/components/Header/index.tsx
@@ -11,7 +11,7 @@ import { useDarkModeManager } from '../../state/user/hooks'
 import { useNativeCurrencyBalance, useTokenBalance } from '../../state/wallet/hooks'
 import { ReactComponent as GasInfoSvg } from '../../assets/svg/gas-info.svg'
 
-import Settings from '../Settings'
+import { Settings } from '../Settings'
 
 import Row, { RowFixed, RowFlat } from '../Row'
 import Web3Status from '../Web3Status'
